requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
decorator==4.4.2
Faker==4.1.2
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-testrail==2.8.3
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
retry==0.9.2
selenium==3.141.0
singledispatch==3.4.0.3
six==1.15.0
testrail==0.3.13
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
webdriver-manager==3.2.1
zipp==3.1.0
